+++ customer_portal/src/epilot/models/operations/getcustomerbalance.py
@@ -9,7 +9,7 @@ from typing import Optional
 
 @dataclasses.dataclass
 class GetCustomerBalanceSecurity:
-    epilot_auth: str = dataclasses.field(metadata={'security': { 'scheme': True, 'type': 'http', 'sub_type': 'bearer', 'field_name': 'Authorization' }})
+    portal_auth: str = dataclasses.field(metadata={'security': { 'scheme': True, 'type': 'http', 'sub_type': 'bearer', 'field_name': 'Authorization' }})